// src/auth/enums/role.enum.ts
export enum UserRole {
  SYSTEM_ADMINISTRATOR = 'SYSTEM_ADMINISTRATOR',
  MANAGER = 'MANAGER',
  SUPERVISOR = 'SUPERVISOR',
  OHS_PRACTITIONER = 'OHS_PRACTITIONER',
  OHS_NATIONAL_OFFICE = 'OHS_NATIONAL_OFFICE',
  FIRST_AIDER = 'FIRST_AIDER',
  HR = 'HR',
  FINANCE_OFFICIAL = 'FINANCE_OFFICIAL',
  EMPLOYEE = 'EMPLOYEE',
  PSSC_COORDINATOR = 'PSSC_COORDINATOR',
  DEPUTY_DIRECTOR = 'DEPUTY_DIRECTOR',
  CHIEF_DIRECTOR = 'CHIEF_DIRECTOR',
  FACILITIES_COORDINATOR = 'FACILITIES_COORDINATOR',
}